Still working on refactoring the read/write of Disk to utilize the Sector struct everywhere.

+++ src/main.rs
@@ -11,7 +11,7 @@ struct Args {
 
     /// Disk size in bytes
     #[clap(short, long)]
-    size: u64,
+    size: usize,
 
     /// Debug flag
     #[clap(short, long)]
